This is a simple application, Makes mouse move
This is a simple application I made for the developers of the 
Leap Motion who still want to use it as a "Mouse". The interface is a 
little bit glitchy. Point on the screen were you want the cursor to
go. The best way to click is to use one finger to point at the thing
you want to click on then quickly with the other hand point anywhere on
the screen and then remove that hand. to scroll down show 3 fingers
to scroll up show 4 fingers. To exit the application show 8 fingers.
The best way to show 8 fingers is 4 on each hand because the Leap
recognizes this the quickest. I am not responsible of any damage done
to your computer by this application(the most common damage that is
done is the leap data glitching out and dragging stuff around, so far
in my testing this aplication has done nothing more than drag things 
around). You can uncomment a section of code out to enable "Jab Clicking"
aka point a "jab" at the thing on the screen to click. This was disabled 
because the mouse tended to move during this action.